+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Refactor] Replace `define_method` with `class_eval` in some locations.
 - [Fix] Fix a case where internal Idle job scheduling would go via the consumption flow.
 - [Fix] Make the Iterator `#stop_partition` work with karafka-rdkafka `0.14.6`.
+- [Ignore] option --include-consumer-groups not working as intended after removal of "thor"
 
 ## 2.2.14 (2023-12-07)
 - **[Feature]** Provide `Karafka::Admin#delete_consumer_group` and `Karafka::Admin#seek_consumer_group`.